Source/requirements.txt
comtypes==1.4.4
pycaw==20240210
pyserial==3.5
